pyproject.toml
[project]
name = "chestguard"
version = "2.0.0"
description = "ChestGuard NeuralScan - AI-Powered Chest X-Ray Diagnostic Intelligence Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "fpdf>=1.7.2",
    "langchain>=0.2.0",
    "langchain-groq>=0.1.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
    "scikit-image>=0.20.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchxrayvision>=1.4.0",
    "langchain-core>=1.2.16",
]

[project.scripts]
chestguard = "server:create_app"
